fix: [os-32] rename ProductPageContainer
